Move bot (Handler) into its own module
