+++ src/main.rs
@@ -1,57 +1,21 @@
 use std::env;
-use std::sync::Arc;
-use tokio::time::Duration;
 use tokio::fs;
 use dotenv::dotenv;
 use serenity::all::GatewayIntents;
 use serenity::prelude::*;
-use serenity::{all::Ready, async_trait};
 use sqlx::sqlite::{SqliteConnectOptions, SqlitePool};
 
 mod config;
-use config::{load_config, write_config, Config};
 mod task;
-use task::Task;
+mod worker;
 mod api;
-use api::{Api, Options};
+mod bot;
+
+use config::{load_config, write_config, Config};
+use task::Task;
+use bot::Handler;
 
 const DEFAULT_PATH: &str = "./config.toml";
-
-struct Handler {
-    client_id: String,
-    client_secret: String,
-    tasks: Vec<Task>,
-    _conn: SqlitePool,
-}
-
-#[async_trait]
-impl EventHandler for Handler {
-    async fn ready(&self, ctx: Context, ready: Ready) {
-        println!("{} is connected!", ready.user.name);
-
-        let ctx = Arc::new(ctx);
-        let tasks = self.tasks.clone();
-        let api = Api::new(self.client_id.clone(), self.client_secret.clone());
-
-        tokio::spawn(async move {
-            let mut tasks = tasks;
-            let mut api = api;
-            loop {
-                check(Arc::clone(&ctx), &mut tasks, &mut api).await;
-                tokio::time::sleep(Duration::from_secs(400)).await;
-            }
-        });
-    }
-}
-
-async fn check(_ctx: Arc<Context>, tasks: &mut Vec<Task>, sirius: &mut Api) {
-    println!("Tasks: {:#?}", tasks);
-    let opts = Options {
-        ..Options::default()
-    };
-    let res = sirius.course_events("BI-LA1.21".into(), opts).await;
-    println!("Response: {:#?}", res);
-}
 
 #[tokio::main]
 async fn main() {
@@ -104,12 +68,12 @@ async fn main() {
 
     let intents = GatewayIntents::GUILD_MESSAGES;
     let mut client = Client::builder(&token, intents)
-        .event_handler(Handler {
+        .event_handler(Handler::new(
             client_id,
             client_secret,
             tasks,
-            _conn: db_pool,
-        })
+            db_pool
+        ))
         .await
         .expect("Error while creating the client!");
